test: organize test files by function
