+++ test/helpers/assert.ts
@@ -1,0 +1,1 @@
+export function assert<T, U extends T>(t: { pass: any }) { t.pass(); }
