+++ test/objects/DiffKeys.test.ts
@@ -1,0 +1,15 @@
+import test from 'ava';
+import { assert } from '../helpers/assert';
+
+import { DiffKeys } from '../../src';
+
+test('Can get all keys that are different between objects', t => {
+    type a = { x: number, y: string };
+    type b = { y: string, z: number };
+
+    type gotA = DiffKeys<a, b>;
+    type gotB = DiffKeys<b, a>;
+
+    assert<gotA, 'x'>(t);
+    assert<gotB, 'z'>(t);
+});
